Add a --type-set option which mimics the corresponding option in ack.
Note that due to the way option processing currently works, --type-set
<new_type> can be used with e.g.: --type php, but NOT with --php
(because processing of the latter type of option happens at an earlier
stage than the processing of the --type-set option that creates the new
file type). This is true even with commit
ae750d128ac57751dfc35fc58fd70b46aa67e741i, which ~will~ help if the file
type already exists.